requirements/base.txt
pytz==2023.3  # https://github.com/stub42/pytz
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==9.5.0  # https://github.com/python-pillow/Pillow
rcssmin==1.1.1  # https://github.com/ndparker/rcssmin
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.4.0  # https://github.com/evansd/whitenoise
redis==4.5.4  # https://github.com/redis/redis-py
hiredis==2.2.2  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django==4.1.8  # pyup: < 4.2  # https://www.djangoproject.com/
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.54.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-compressor==4.3.1  # https://github.com/django-compressor/django-compressor
django-redis==5.2.0  # https://github.com/jazzband/django-redis
djangorestframework==3.14.0
django-cors-headers==3.14.0
openai==0.27.4
djangorestframework_simplejwt==5.2.2
google-api-python-client==2.86.0
stripe==5.4.0
